Split middleware into a subpackage
